Updated displayed information
Updated the numbers on the roi and totals table to both be to the nearest hundredths decimal place. I also updated the path table to only display the file name after Sales Record to shorten the name of the file.

Continued to test functions and found a bug within login where if you enter a username and fill in the first password space it will allow you to sign up.