Updated TableSubject
Began with removing the suppression warning coming from notifyObservers() in TableSubject. With that removal I was able to find the bug of our getValueAt() function within roiTableWriter. Since we had multiple data types such as String, boolean and Double, it would return a mismatch error. After which I traced it and found its corresponding function getRowOrderObject() was the other issue. See new code for changes